feat: Add sad path
